+++ src/main/java/dev/mark/jewelsstorebackend/products/ProductDTO.java
@@ -3,6 +3,6 @@ package dev.mark.jewelsstorebackend.products;
 public class ProductDTO {
     public String productName;
     public String productDescription;
-    public float price;
+    public Long price;
     public Long categoryId;
 }
